Fix potential race condition in initialize
The delay before calling doEverything() averted some
race conditions, but assigning the onClick action to
the memoryCardDeck was still done outside of that
delayed action. It's conceivable that that might run
before the DOM was fully loaded.

Also fixed proper resetting of the move counter when
the repeat button is clicked (it used to reset the
move counter but did not immediately display the new
zero value after reset).